+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 # Change Log
 
-## 0.4.1 - 2020-02-26
+## 0.4.2 - 2020-03-10
+
 ### Features
-- 아이나비 로고 이미지 변경
+- 특정 영역 전체가 보이는 카메라 위치 정보를 반환하는 API 추가 
+  - `InaviMap#getCameraFitBounds`
+
+- 지도에 등록된 모든 셰이프를 지도에서 제거하는 API 추가
+  - `InaviMap#clearShapes`
+
+### Bug Fixes
+- 지도 초기화 전에 일부 셰이프 객체 생성 시 지도에 비정상 표출되는 오류 수정
+
+## 0.4.1 - 2020-02-26
+
+### Features
+- 지도 로고 이미지 변경
 
 ### Bug Fixes
 - Proguard 난독화 적용 시 비정상 종료되는 오류 수정
@@ -38,11 +51,11 @@
 ## 0.3.1 - 2019-12-18
 
 ### Features
- - 지도 이동 영역을 제한하는 I/F 추가
+ - 지도 이동 영역 제한 API 추가
    - `InaviMap#setConstraintBounds`, `InaviMap#getConstraintBounds`
- - 로고 클릭 이벤트 활성화 여부 설정 I/F 추가
+ - 로고 클릭 이벤트 활성화 여부 설정 API 추가
    - `UiSettings#setLogoClickEnabled`, `UiSettings#isLogoClickEnabled`
- - 오픈소스 라이선스, 법적 공지 Activity 호출 Intent I/F 추가
+ - 오픈소스 라이선스, 법적 공지 Activity Intent 반환 API 추가
    - `InaviMapSdk#getIntentLicenseActivity`, `InaviMapSdk#getIntentLegalNoticeActivity`
  - 마커 아이콘과 타이틀 사이의 여백을 설정하는 기능 추가
    - `InvMarker#setTitleMargin`, `InvMarker#getTitleMargin`